requirements.txt
fastapi
uvicorn[standard]
pydantic
python-jose[cryptography]
pymysql
sqlalchemy
pydantic[email]


--- IGNORE ---
python -m uvicorn app.main:app --reload